[major] move the components in example/Transducer_Examples to components in the core code base.
