+++ sequence/components/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ['beam_splitter', 'bsm', 'detector', 'interferometer', 'light_source', 'memory', 'optical_channel', 'photon',
-           'spdc_lens', 'switch', 'circuit']
+           'spdc_lens', 'switch', 'circuit', 'transmon', 'transducer']
 
 def __dir__():
     return sorted(__all__)
