+++ sequence/components/beam_splitter.py
@@ -118,3 +118,35 @@ class FockBeamSplitter(Entity):
             detector_num = self.get_generator().choice([0, 1], p=[prob_0, 1-prob_0])
             self.most_recent_time = self.timeline.now()
             self._receivers[detector_num].get()
+
+
+
+class FockBeamSplitter2(Entity):
+    """Class modeling a Fock beam splitter.
+
+    A Fock beam splitter can send a single photon randomly in one of its ports.
+
+    See https://arxiv.org/abs/2411.11377, Simulation of Quantum Transduction Strategies for Quantum Networks
+
+    """
+    def __init__(self, name: str, owner: "Node", timeline: "Timeline", efficiency: int, photon_counter: int, src_list: List[str]):
+        Entity.__init__(self, name, timeline)
+        self.name = name
+        self.owner = owner
+        self.timeline = timeline
+        self.efficiency = efficiency
+        self.photon_counter = photon_counter
+        
+    def init(self):
+        assert len(self._receivers) == 2
+
+    def receive_photon_from_scr(self, photon: "Photon", source: List[str]) -> None:
+        self.photon_counter += 1
+
+    def add_output(self, outputs: List):
+        for i in outputs:
+            self.add_receiver(i)
+
+    def send_photon(self, receiver: "Entity", photon: "Photon") -> None:
+        receiver.get(self.name, photon)
+
